App: Move ValidationPanel to its own file.
