+++ packages/openneuro-app/src/scripts/datalad/validation/validation-status.jsx
@@ -1,45 +1,8 @@
 import React from 'react'
 import PropTypes from 'prop-types'
-import { Panel, PanelGroup } from 'react-bootstrap'
 import pluralize from 'pluralize'
+import ValidationPanel from './validation-panel.jsx'
 import Results from '../../upload/upload.validation-results.jsx'
-
-class ValidationPanel extends React.Component {
-  constructor(props) {
-    super(props)
-    this.state = {
-      activeKey: null,
-    }
-    this.togglePanel = this.togglePanel.bind(this)
-  }
-
-  togglePanel() {
-    if (this.state.activeKey === '1') {
-      this.setState({ activeKey: '2' })
-    } else if (this.state.activeKey === '2') {
-      this.setState({ activeKey: '1' })
-    }
-  }
-
-  render() {
-    return (
-      <PanelGroup
-        accordion
-        className="validation-wrap"
-        activeKey={this.state.activeKey}
-        onSelect={this.togglePanel}>
-        <Panel className="status" header={this.props.heading} eventKey="1">
-          {this.props.children}
-        </Panel>
-      </PanelGroup>
-    )
-  }
-}
-
-ValidationPanel.propTypes = {
-  heading: PropTypes.object,
-  children: PropTypes.array,
-}
 
 /**
  * These can't be React components due to legacy react-bootstrap
